Consolidate .netcoreapp.cs files because System.Runtime.* projects is no longer cross-compiled
