+++ src/libraries/System.Runtime/tests/System.Runtime.Tests.csproj
@@ -259,7 +259,7 @@
     <Compile Include="$(CommonPath)..\tests\System\RealFormatterTestsBase.netcoreapp.cs" Link="System\RealFormatterTestsBase.netcoreapp.cs" />
     <Compile Include="System\RealFormatterTests.cs" />
     <Compile Include="$(CommonPath)..\tests\System\RealParserTestsBase.netcoreapp.cs" Link="System\RealParserTestsBase.netcoreapp.cs" />
-    <Compile Include="System\RealParserTests.netcoreapp.cs" />
+    <Compile Include="System\RealParserTests.cs" />
   </ItemGroup>
   <ItemGroup Condition="'$(TargetsNetCoreApp)' != 'true'">
     <Compile Include="System\StringSplitExtensions.cs" />
